fix: v4.2.15 breaks building docs
In v4.2.15 we have added a feature to include scripts and styles only in
target files. In order to do so we are looping through `this.includeScript`
and `this.includeCss`. However we are not checking whether these are
defined.

So, for users who don't want to include any script/style above mentioned
variables are undefined for then, hence cause the build to fail because
undefined is not iterable.

This PR fixes this issue, as now we are checking whether these variables
are defined.

fixes: #259