Fix highlighting issue for links having spaces.
When there is a space in the url then `document.location.href`` returns a url that has `%20` in it. Now in `highlightActiveLinkInSidebar` we are looking for anchor element having href similar to `document.location.href`.

Currently we are not decoding the URL meaning `%20` is not converted into space (` `). In anchor tab the href has space instead of `%20`, so to fix this issue we are using `decodeURI` to decode the url returned by `document.location.href`

fixes: #283